# inventory-svc/api-spec.yaml
swagger: "2.0"
info:
  title: inventory-service-api-spec
  version: "1.0.0"
tags:
- name: "item"
  description: "Endpoints for manipulating items"
paths:
  /items:
    get:
      tags:
        - "item"
      operationId: getAllitems
      summary: Retrieves all items
      produces:
      - application/json
      responses:
        200:
          description: |-
            200 response
          examples:
            application/json: |-
              [
                {
                  "name": "item1",
                  "price": 1.11
                  "id": 1
                },
                {
                  "name": "item2",
                  "price": 2.22
                  "id": 2
                }
              ]

    post:
      tags:
      - "item"
      operationId: additem
      summary: Adds a new item
      consumes:
        - application/json
      parameters:
        - in: body
          name: item
          schema:
            type: object
            required:
              - name
              - price
            properties:
              name:
                type: string
              price:
                type: number
      produces:
        - application/json
      responses:
        201:
          description: Created Item
          schema:
            type: object
            properties:
              id:
                type: integer
                description: Item Id
              name:
                type: string
                description: Item name
              price:
                type: number
                description: Item price

    delete:
      tags:
       - "item"
      operationId: deleteitem
      summary: Deletes an item
      consumes:
      - application/json
      parameters:
      - in: body
        name: item
        schema:
          type: object
          required:
          - name
          - id
          properties:
            name:
              type: string
            id:
              type: integer
      responses:
        204:
          description: Item Deleted

host: 35.234.149.119/inventory
schemes:
- http
consumes:
- application/json